+++ lib/src/buttons/subtitle_button.dart
@@ -11,7 +11,6 @@ class SubtitleButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.all(8.0),
       child: new Cover(
         icon: Icons.closed_caption,
         onTap: () {