+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [2.1.9] - Mon November 17, 2020
+
+* Perfect seek buttons 
+
+## [2.1.8] - Mon November 16, 2020
+
+* Add custom seek
+* Add seek gestures and remove buttons
+* Remove bool showSeekButtons controller member data
+
+## [2.1.7] - Mon October 12, 2020
+
+* Add next, previous, setIndex, pause, play, callbacks
+* Correct was called on null error
+
 ## [2.1.6] - Mon August 02, 2020
 
 * Update out of date packages